Use PlutusTx.Ord.< instead of Prelude.<. Move Min, Max to Semigroup.
