+++ plutus-tx/src/PlutusTx/List.hs
@@ -22,9 +22,10 @@ module PlutusTx.List (
 import           PlutusTx.Bool     ((||))
 import qualified PlutusTx.Builtins as Builtins
 import           PlutusTx.Eq       (Eq, (==))
+import           PlutusTx.Ord      ((<))
 import           PlutusTx.Trace    (traceError)
 import           Prelude           hiding (Eq (..), all, any, elem, filter, foldl, foldr, head, length, map, null,
-                                    reverse, tail, take, zip, (!!), (&&), (++), (||))
+                                    reverse, tail, take, zip, (!!), (&&), (++), (<), (||))
 
 {- HLINT ignore -}
